'Resolving' various cast warnings from MSVC.  We have problems,  though,  where ERR (-1) is cast to an unsigned type,  and with MS's insane decision to make wint_t 16 bits,  effectively locking off anything but the BMP for some functions.

+++ pdcurses/util.c
@@ -314,7 +314,7 @@ size_t PDC_mbstowcs(wchar_t *dest, const char *src, size_t n)
         int retval = PDC_mbtowc(dest + i, src, len);
 
         if (retval < 1)
-            return -1;
+            return (size_t)-1;
 
         src += retval;
         len -= retval;
